+++ darts/dataprocessing/pipeline.py
@@ -124,6 +124,18 @@ class Pipeline:
         for transformer in reversed(self._transformers):
             data = transformer.inverse_transform(data)
         return data
+    
+    def invertible(self) -> bool:
+        """
+        Returns whether the pipeline is invertible or not.
+        A pipeline is invertible if all transformers in the pipeline are themselves invertible.
+
+        Returns
+        -------
+        bool
+            True if the pipeline is invertible, False otherwise
+        """
+        return self._invertible
 
     def __getitem__(self, key: Union[int, slice]) -> 'Pipeline':
         """